pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ptz-autotrack"
version = "0.1.0"
description = "Production-grade real-time person auto-tracking for Dahua PTZ cameras"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "carrot"}
]
dependencies = [
  "opencv-python-headless>=4.10.0",
  "ultralytics>=8.3.0",
  "requests>=2.32.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "numpy>=1.26.0",
  "PyYAML>=6.0.1",
  "structlog>=24.4.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "prometheus-client>=0.20.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0"
]

[project.scripts]
ptz-autotrack = "app.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
